Session expiration

+++ webapp.py
@@ -194,7 +194,7 @@ def get_google_provider_cfg():
 @app.route('/sbhs')
 @app.route('/sbhs/<space_id>')
 def render_main_page(space_id = None):
-    if space_id != None:
+    if space_id != None and session_valid():
         if 'logged' not in session or session['logged'] == False:
             session['invite'] = space_id
             return redirect(url_for('render_login'))
@@ -210,7 +210,7 @@ def render_main_page(space_id = None):
 
 @app.route('/logout', methods=['GET', 'POST'])
 def logout():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         session['logged'] = False # Prevents browsers from using cached session data to log in. NOTE: We use server-side sessions now
         session.clear()
         return Response(dumps({'success': 'true'}), mimetype='application/json')
@@ -219,7 +219,7 @@ def logout():
 
 @app.route('/list_spaces', methods=['GET', 'POST'])
 def list_spaces():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         user_spaces = collection_users.find_one({"_id": session['unique_id']})['joined']
         space_list = []
         for space_item in user_spaces:
@@ -235,7 +235,7 @@ def list_spaces():
 
 @app.route('/user_spaces', methods=['GET', 'POST'])
 def user_spaces():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         user_spaces = collection_users.find_one({"_id": session['unique_id']})['joined']
         space_list = []
         for space_item in user_spaces:
@@ -250,7 +250,7 @@ def user_spaces():
 
 @app.route('/space', methods=['GET', 'POST'])
 def render_space():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         space = collection_spaces.find_one({'_id': ObjectId(request.json['space_id'])})
         rooms_and_sections = dumps([list(collection_rooms.find({'space': request.json['space_id']}).sort('order', 1)), list(collection_sections.find({'space': request.json['space_id']}).sort('order', 1)), list(collection_users.find({'joined': {'$in': [request.json['space_id']]}})), list(collection_spaces.find({'_id': ObjectId(request.json['space_id'])}))])
         if session['unique_id'] in space['admins'] or session['admin']:
@@ -268,7 +268,7 @@ def render_space():
 
 @app.route('/leave_space', methods=['GET', 'POST'])
 def leave_space():
-    if request.method == 'POST' and not space_owner():
+    if request.method == 'POST' and session_valid() and not space_owner():
         joined = collection_users.find_one({"_id": session['unique_id']})['joined']
         joined.remove(session['current_space'])
         collection_users.update_one({"_id": session['unique_id']}, {"$set": {"joined": joined}})
@@ -284,13 +284,13 @@ def leave_space():
 
 @app.route('/chat_history', methods=['GET', 'POST'])
 def chat_history():
-    if request.method == 'POST' and space_member():
+    if request.method == 'POST' and session_valid() and space_member():
         chat_history = dumps(list(collection_messages.find({'room': request.json['room_id']}).sort('_id', pymongo.DESCENDING).skip(int(request.json['i'])).limit(50)))
         return Response(chat_history, mimetype='application/json')
 
 @app.route('/email_history', methods=['GET', 'POST']) #return only emails users can see, and check if space admin
 def email_history():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         email_history = collection_emails.find({'room': request.json['room_id']}).sort('_id', pymongo.DESCENDING).skip(int(request.json['i'])).limit(50)
         email_list = []
         for email in email_history:
@@ -300,7 +300,7 @@ def email_history():
 
 @app.route('/send_email', methods=['GET', 'POST'])
 def send_email():
-    if request.method == 'POST' and (space_admin() or session['admin']):
+    if request.method == 'POST' and session_valid() and (space_admin() or session['admin']):
         sender_email = 'sbhs.platform.test@gmail.com'
         password = os.environ['EMAIL_ACCESS_PASSWORD']
         message = MIMEMultipart('alternative')
@@ -334,7 +334,7 @@ def send_email():
 
 @app.route('/delete_room', methods=['GET', 'POST'])
 def delete_room():
-	if request.method == 'POST' and (space_admin() or session['admin']):
+	if request.method == 'POST' and session_valid() and (space_admin() or session['admin']):
 		room_count = collection_rooms.count_documents({'space': session['current_space'], 'section': 'special'}) - collection_rooms.count_documents({'space': session['current_space']})
 		if room_count > 1:
 			collection_rooms.delete_one({'_id': ObjectId(request.json['room_id'])})
@@ -353,7 +353,7 @@ def delete_room():
 
 @app.route('/create_room', methods=['GET', 'POST'])
 def create_room():
-	if request.method == 'POST' and (space_admin() or session['admin']):
+	if request.method == 'POST' and session_valid() and (space_admin() or session['admin']):
 		room_id = ObjectId()
 		room_list = list(collection_rooms.find({'space': session['current_space'], 'section': request.json['section_id']}))
 		room = {'_id': room_id, 'space': session['current_space'], 'section': request.json['section_id'], 'name': request.json['room_name'], 'order': len(room_list) + 1}
@@ -368,7 +368,7 @@ def create_room():
 
 @app.route('/create_section', methods=['GET', 'POST'])
 def create_section():
-	if request.method == 'POST' and (space_admin() or session['admin']):
+	if request.method == 'POST' and session_valid() and (space_admin() or session['admin']):
 		section_id = ObjectId()
 		section_list = list(collection_sections.find({'space': session['current_space']}))
 		section = {'_id': section_id, 'space': session['current_space'], 'name': request.json['section_name'], 'order': len(section_list) + 1}
@@ -382,7 +382,7 @@ def create_section():
 
 @app.route('/delete_section', methods=['GET', 'POST'])
 def delete_section():
-    if request.method == 'POST' and (space_admin() or session['admin']):
+    if request.method == 'POST' and session_valid() and (space_admin() or session['admin']):
         section_count = collection_sections.count_documents({'space': session['current_space']})
         if section_count > 1:
             collection_sections.delete_one({'_id': ObjectId(request.json['section_id'])})
@@ -407,25 +407,26 @@ def delete_section():
 
 @app.route('/create_space', methods=['GET', 'POST']) #Check if space with name already exists...
 def create_space():
-    user = collection_users.find_one({"_id": session['unique_id']})
-    if request.method == 'POST' and user['owns'] < 3:
-        space_id = ObjectId()
-        room_id = ObjectId()
-        email_room_id = ObjectId()
-        section_id = ObjectId()
-        room = {'_id': room_id, 'space': str(space_id), 'section': str(section_id), 'name': 'general', 'order': 1}
-        special_rooms = {'_id': email_room_id, 'space': str(space_id), 'section': 'special', 'name': 'Email', 'order': 1}
-        section = {'_id': section_id, 'space': str(space_id), 'name': 'discussion', 'order': 1}
-        image = request.json['space_image']
-        if (requests.get(image, stream = True).headers['Content-length'] > 6000000 or not requests.head(image).headers["content-type"] in ("image/png", "image/jpeg", "image/jpg", "image/gif", "image/avif", "image/webp", "image/svg")):
-            image = '/static/images/Space.jpeg'
-        collection_spaces.insert_one({'_id': space_id, 'name': request.json['space_name'], 'picture': image, 'admins': [session['unique_id']], 'members': [[session['unique_id'], session['users_name']]]})
-        collection_rooms.insert_many([room, special_rooms])
-        collection_sections.insert_one(section)        
-        joined = user['joined']
-        joined.append(str(space_id))
-        collection_users.find_one_and_update({"_id": session['unique_id']}, {'$set': {'joined': joined, 'owns': user['owns'] + 1}})
-        return Response(dumps({'space_id': str(space_id)}), mimetype='application/json')
+    if session_valid():
+        user = collection_users.find_one({"_id": session['unique_id']})
+        if request.method == 'POST' and user['owns'] < 3:
+            space_id = ObjectId()
+            room_id = ObjectId()
+            email_room_id = ObjectId()
+            section_id = ObjectId()
+            room = {'_id': room_id, 'space': str(space_id), 'section': str(section_id), 'name': 'general', 'order': 1}
+            special_rooms = {'_id': email_room_id, 'space': str(space_id), 'section': 'special', 'name': 'Email', 'order': 1}
+            section = {'_id': section_id, 'space': str(space_id), 'name': 'discussion', 'order': 1}
+            image = request.json['space_image']
+            if (requests.get(image, stream = True).headers['Content-length'] > 6000000 or not requests.head(image).headers["content-type"] in ("image/png", "image/jpeg", "image/jpg", "image/gif", "image/avif", "image/webp", "image/svg")):
+                image = '/static/images/Space.jpeg'
+            collection_spaces.insert_one({'_id': space_id, 'name': request.json['space_name'], 'picture': image, 'admins': [session['unique_id']], 'members': [[session['unique_id'], session['users_name']]]})
+            collection_rooms.insert_many([room, special_rooms])
+            collection_sections.insert_one(section)        
+            joined = user['joined']
+            joined.append(str(space_id))
+            collection_users.find_one_and_update({"_id": session['unique_id']}, {'$set': {'joined': joined, 'owns': user['owns'] + 1}})
+            return Response(dumps({'space_id': str(space_id)}), mimetype='application/json')
     return Response(dumps({'success': 'false'}), mimetype='application/json')
 
 # Adds space to user's list of joined spaces in MongoDB.
@@ -434,7 +435,7 @@ def create_space():
 
 @app.route('/delete_space', methods=['GET', 'POST'])
 def delete_space():
-    if request.method == 'POST' and (space_owner() or session['admin']):
+    if request.method == 'POST' and session_valid() and (space_owner() or session['admin']):
         space = collection_spaces.find_one({'_id': ObjectId(session['current_space'])})
         collection_spaces.delete_one({'_id': ObjectId(session['current_space'])})
         collection_users.find_one_and_update({"_id": space['admins'][0]}, {'$inc': {'owns': -1}})
@@ -448,7 +449,7 @@ def delete_space():
 
 @app.route('/join_space', methods=['GET', 'POST'])
 def join_space():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         joined = collection_users.find_one({"_id": session['unique_id']})['joined']
         space = dumps(collection_spaces.find_one({'_id': ObjectId(request.json['space_id'])}))
         if request.json['space_id'] not in joined:
@@ -464,7 +465,7 @@ def join_space():
 
 @app.route('/delete_message', methods=['GET', 'POST']) #space admin and message in space
 def delete_message():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         deleted_message = collection_messages.find_one({'_id': ObjectId(request.json['message_id'])})
         if deleted_message['user_id'] == session['unique_id'] or space_admin() or session['admin']:
             deleted_email = collection_messages.find({'room': request.json['room_id'], 'email': deleted_message['email']}).sort('_id', pymongo.DESCENDING)
@@ -485,7 +486,7 @@ def delete_message():
 
 @app.route('/report_message', methods=['GET', 'POST'])
 def report_message():
-    if request.method == 'POST' and space_member():
+    if request.method == 'POST' and session_valid() and space_member():
         reported_message = collection_messages.find_one({'_id': ObjectId(request.json['message_id'])})
         if collection_logs.count_documents({'details': reported_message}) == 0:
             collection_logs.insert_one({'name': session['users_name'], 'user_id': session['unique_id'], 'email': session['users_email'], 'action': 'reported message', 'by': reported_message['name'], 'by_email': reported_message['email'], 'in': session['current_space_name'], 'details': reported_message, 'datetime': datetime.now().isoformat() + 'Z'})
@@ -500,7 +501,7 @@ def report_message():
 
 @app.route('/open_member_profile', methods=['GET', 'POST'])
 def member_profile():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         member = collection_users.find_one({'_id': request.json['user_id']})
         queried_spaces = []
         names_list = []
@@ -521,7 +522,7 @@ def member_profile():
 
 @app.route('/profile', methods=['GET', 'POST'])
 def profile():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         data = collection_users.find_one({'_id': session['unique_id']})
         return Response(dumps(data), mimetype='application/json')
     else:
@@ -529,7 +530,7 @@ def profile():
 
 @app.route('/sorted_spaces', methods=['GET', 'POST'])
 def sorted_spaces():
-    if request.method == 'POST':
+    if request.method == 'POST' and session_valid():
         collection_users.find_one_and_update({"_id": session['unique_id']}, {'$set': {'joined': request.json['space_list']}})
         return Response(dumps({'success': 'true'}), mimetype='application/json')
     else:
@@ -537,19 +538,19 @@ def sorted_spaces():
 
 @app.route('/server_logs', methods=['GET', 'POST'])
 def server_logs():
-    if request.method == 'POST' and session['admin']:
+    if request.method == 'POST' and session_valid() and session['admin']:
         logs = dumps(list(collection_logs.find().sort('_id', pymongo.DESCENDING).skip(int(request.json['i'])).limit(1000)))
         return Response(logs, mimetype='application/json')
 
 @app.route('/server_users', methods=['GET', 'POST'])
 def server_users():
-    if request.method == 'POST' and session['admin']:
+    if request.method == 'POST' and session_valid() and session['admin']:
         users = dumps(list(collection_users.find().sort('_id', pymongo.DESCENDING)))
         return Response(users, mimetype='application/json')
 
 @app.route('/change_user_status', methods=['GET', 'POST'])
 def change_user_status():
-    if request.method == 'POST' and session['admin']:
+    if request.method == 'POST' and session_valid() and session['admin']:
         if request.json['status'] in {'banned', 'user', 'admin'}:
             collection_users.find_one_and_update({"_id": request.json['user_id']}, {'$set': {'status': request.json['status']}})
             return Response(dumps({'success': 'true'}), mimetype='application/json')
@@ -560,7 +561,7 @@ def change_user_status():
 
 @socketio.on('join_room')
 def change_room(data):
-    if space_member() and valid_room(data['room_id']):
+    if session_valid() and (space_member() and valid_room(data['room_id'])):
         try: 
             leave_room(data['old_room'])
         except:
@@ -571,14 +572,14 @@ def change_room(data):
 
 @socketio.on('is_typing')
 def is_typing(data):
-    if space_member() and valid_room(data['room_id']):
+    if session_valid() and (space_member() and valid_room(data['room_id'])):
 	    socketio.emit('is_typing', data, room = data['room_id'])
 
 # When user stops typing, notify all users in that room.
 
 @socketio.on('stopped_typing')
 def stopped_typing(data):
-    if space_member() and valid_room(data['room_id']):
+    if session_valid() and (space_member() and valid_room(data['room_id'])):
 	    socketio.emit('stopped_typing', data, room = data['room_id'])
 
 # When a message is sent, verify and store it in MongoDB.
@@ -586,7 +587,7 @@ def stopped_typing(data):
 
 @socketio.on('send_message')
 def send_message(data):
-    if space_member() and valid_room(data['room_id']):
+    if session_valid() and (space_member() and valid_room(data['room_id'])):
         utc_dt = datetime.now().isoformat() + 'Z'
         data['datetime'] = utc_dt
         data['message'] = re.sub('\\\n\\n\\\n+', '\\n\\n', data['message'])
@@ -611,7 +612,7 @@ def send_message(data):
 
 @socketio.on('created_room')
 def created_room(data):
-    if space_admin() or session['admin']:
+    if session_valid() and (space_admin() or session['admin']):
         for room in room_list():
             socketio.emit('created_room', data, room = room)
 
@@ -620,7 +621,7 @@ def created_room(data):
 
 @socketio.on('deleted_room')
 def deleted_room(data):
-    if space_admin() or session['admin']:
+    if session_valid() and (space_admin() or session['admin']):
         for room in room_list():
             socketio.emit('deleted_room', data, room = room)
 
@@ -635,7 +636,7 @@ def deleted_space():
 
 @socketio.on('created_section')
 def created_section(data):
-    if space_admin() or session['admin']:
+    if session_valid() and (space_admin() or session['admin']):
         for room in room_list():
             socketio.emit('created_section', data, room = room)
 
@@ -644,7 +645,7 @@ def created_section(data):
 
 @socketio.on('deleted_section')
 def created_section(data):
-    if space_admin() or session['admin']:
+    if session_valid() and (space_admin() or session['admin']):
         for room in room_list():
             socketio.emit('deleted_section', data, room = room)
 
@@ -653,7 +654,7 @@ def created_section(data):
 
 @socketio.on('deleted_message')
 def deleted_message(data):
-    if space_admin() or session['admin'] or session['unique_id'] == collection_messages.find_one({'_id': ObjectId(data['message_id'])})['user_id']:
+    if session_valid() and (space_admin() or session['admin'] or session['unique_id'] == collection_messages.find_one({'_id': ObjectId(data['message_id'])})['user_id']):
         socketio.emit('deleted_message', data, room = data['room_id'])
 
 # When a message is edited, update the message in MongoDB and
@@ -661,7 +662,7 @@ def deleted_message(data):
 
 @socketio.on('edited_message')
 def edited_message(data):
-    if space_admin() or session['admin'] or session['unique_id'] == collection_messages.find_one({'_id': ObjectId(data['message_id'])})['user_id']:
+    if session_valid() and (space_admin() or session['admin'] or session['unique_id'] == collection_messages.find_one({'_id': ObjectId(data['message_id'])})['user_id']):
         collection_messages.find_one_and_update({"_id": ObjectId(data['message_id'])}, {'$set': {'message': data['edit']}})
         socketio.emit('edited_message', data, room = data['room_id'])
 
@@ -669,7 +670,7 @@ def edited_message(data):
 
 @socketio.on('sorted_sections')
 def sorted_sections(data):
-    if space_admin() or session['admin']:
+    if session_valid() and (space_admin() or session['admin']):
         for section in data['section_list']:
             collection_sections.find_one_and_update({"_id": ObjectId(section)}, {'$set': {'order': data['section_list'].index(section) + 1}})
         for room in room_list():
@@ -679,7 +680,7 @@ def sorted_sections(data):
 
 @socketio.on('sorted_rooms')
 def sorted_rooms(data):
-    if space_admin() or session['admin']:
+    if session_valid() and (space_admin() or session['admin']):
         for section in data['room_group_list']:
             if len(section) > 1:
                 for room in section[1:]:
@@ -705,10 +706,10 @@ def joined_space():
     for room in room_list():
         emit('joined_space', user, room = room, include_self=False)
 
-@app.before_request
-def session_expired():
-    if 'logged' not in session and request.endpoint != 'render_login':
+def session_valid():
+    if 'logged' not in session:
         return redirect(url_for('render_login', error = "Your session expired, please log in again"))
+    return True
 
 def space_admin():
     if session['unique_id'] in collection_spaces.find_one({'_id': ObjectId(session['current_space'])})['admins']:
